refactor: :recycle: use app-sdk syncwebhook
