--- apps/taxes/src/lib/saleor/saleor-app-sdk/index.ts
@@ -1,3 +1,0 @@
-export * from "./saleor-app-sdk-copy";
-export * from "./saleor-sync-webhook";
-export * from "./saleor-webhook";
