--- apps/taxes/src/modules/taxes/defaults.ts
@@ -1,8 +1,0 @@
-import { ResponseTaxPayload } from "./types";
-
-export const defaultTaxesResponse: ResponseTaxPayload = {
-  lines: [],
-  shipping_price_gross_amount: 0,
-  shipping_price_net_amount: 0,
-  shipping_tax_rate: "0",
-};
