--- apps/taxes/src/modules/taxes/tax-webhook.ts
@@ -1,6 +1,0 @@
-import { SaleorSyncWebhook } from "../../lib/saleor/saleor-app-sdk";
-import { ResponseTaxPayload } from "./types";
-export class TaxSaleorSyncWebhook<TPayload = any> extends SaleorSyncWebhook<
-  TPayload,
-  ResponseTaxPayload
-> {}
