+++ apps/taxes/src/modules/taxes/types.ts
@@ -1,22 +1,3 @@
-type LineTaxResponsePayload = {
-  total_gross_amount: number;
-  total_net_amount: number;
-  tax_rate: string;
-};
+import { SyncWebhookResponsesMap } from "@saleor/app-sdk/handlers/next";
 
-export type ResponseTaxPayload = {
-  shipping_price_gross_amount: number;
-  shipping_price_net_amount: number;
-  shipping_tax_rate: string;
-  lines: LineTaxResponsePayload[];
-};
-
-export type FetchTaxesLinePayload = {
-  id: string;
-  quantity: number;
-  taxCode?: string | null;
-  discount: number;
-  chargeTaxes: boolean;
-  unitAmount: number;
-  totalAmount: number;
-};
+export type ResponseTaxPayload = SyncWebhookResponsesMap["ORDER_CALCULATE_TAXES"];
